song-recognition/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Werkzeug==2.3.7
pydub==0.25.1
librosa==0.10.1
numpy==1.24.3
scipy==1.11.4
soundfile==0.12.1
python-dotenv==1.0.0
gunicorn==21.2.0
scikit-learn==1.3.0
matplotlib==3.7.2
pandas==2.0.3
